Conflicts:
	src/libsync/filesystem.cpp
	src/libsync/filesystem.h

+++ src/libsync/propagatedownload.cpp
@@ -24,6 +24,7 @@
 #include "propagatorjobs.h"
 #include "common/checksums.h"
 #include "common/asserts.h"
+#include "common/vfs.h"
 
 #include <QLoggingCategory>
 #include <QNetworkAccessManager>
@@ -66,25 +67,29 @@ QString OWNCLOUDSYNC_EXPORT createDownloadTmpFileName(const QString &previous)
 }
 
 // DOES NOT take ownership of the device.
-GETFileJob::GETFileJob(AccountPtr account, const QString &path, QFile *device,
+GETFileJob::GETFileJob(AccountPtr account, const QString &path, QIODevice *device,
     const QMap<QByteArray, QByteArray> &headers, const QByteArray &expectedEtagForResume,
     quint64 resumeStart, QObject *parent)
     : GETJob(account, path, parent)
     , _device(device)
     , _headers(headers)
     , _expectedEtagForResume(expectedEtagForResume)
+    , _expectedContentLength(-1)
+    , _contentLength(-1)
     , _resumeStart(resumeStart)
     , _hasEmittedFinishedSignal(false)
 {
 }
 
-GETFileJob::GETFileJob(AccountPtr account, const QUrl &url, QFile *device,
+GETFileJob::GETFileJob(AccountPtr account, const QUrl &url, QIODevice *device,
     const QMap<QByteArray, QByteArray> &headers, const QByteArray &expectedEtagForResume,
     quint64 resumeStart, QObject *parent)
     : GETJob(account, url.toEncoded(), parent)
     , _device(device)
     , _headers(headers)
     , _expectedEtagForResume(expectedEtagForResume)
+    , _expectedContentLength(-1)
+    , _contentLength(-1)
     , _resumeStart(resumeStart)
     , _directDownloadUrl(url)
     , _hasEmittedFinishedSignal(false)
@@ -185,6 +190,16 @@ void GETFileJob::slotMetaDataChanged()
         qCWarning(lcGetJob) << "We received a different E-Tag for resuming!"
                             << _expectedEtagForResume << "vs" << _etag;
         _errorString = tr("We received a different E-Tag for resuming. Retrying next time.");
+        _errorStatus = SyncFileItem::NormalError;
+        reply()->abort();
+        return;
+    }
+
+    _contentLength = reply()->header(QNetworkRequest::ContentLengthHeader).toLongLong();
+    if (_expectedContentLength != -1 && _contentLength != _expectedContentLength) {
+        qCWarning(lcGetJob) << "We received a different content length than expected!"
+                            << _expectedContentLength << "vs" << _contentLength;
+        _errorString = tr("We received an unexpected download Content-Length.");
         _errorStatus = SyncFileItem::NormalError;
         reply()->abort();
         return;
@@ -344,17 +359,33 @@ void PropagateDownloadFile::start()
     qCDebug(lcPropagateDownload) << _item->_file << propagator()->_activeJobList.count();
     _stopwatch.start();
 
+    auto &syncOptions = propagator()->syncOptions();
+    auto &vfs = syncOptions._vfs;
+
     // For virtual files just create the file and be done
+    if (_item->_type == ItemTypeVirtualFileDehydration) {
+        _item->_type = ItemTypeVirtualFile;
+        // TODO: Could dehydrate without wiping the file entirely
+        auto fn = propagator()->getFilePath(_item->_file);
+        qCDebug(lcPropagateDownload) << "dehydration: wiping base file" << fn;
+        propagator()->_journal->deleteFileRecord(_item->_file);
+        QFile::remove(fn);
+
+        if (vfs->mode() == Vfs::WithSuffix) {
+            // Normally new suffix-virtual files already have the suffix included in the path
+            // but for dehydrations that isn't the case. Adjust it here.
+            _item->_file.append(vfs->fileSuffix());
+        }
+    }
+    if (vfs->mode() == Vfs::Off && _item->_type == ItemTypeVirtualFile) {
+        qCWarning(lcPropagateDownload) << "ignored virtual file type of" << _item->_file;
+        _item->_type = ItemTypeFile;
+    }
     if (_item->_type == ItemTypeVirtualFile) {
         auto fn = propagator()->getFilePath(_item->_file);
         qCDebug(lcPropagateDownload) << "creating virtual file" << fn;
 
-        // NOTE: Other places might depend on contents of placeholder files (like csync_update)
-        QFile file(fn);
-        file.open(QFile::ReadWrite | QFile::Truncate);
-        file.write(" ");
-        file.close();
-        FileSystem::setModTime(fn, _item->_modtime);
+        vfs->createPlaceholder(propagator()->_localDir, _item);
         updateMetadata(false);
         return;
     }
@@ -631,7 +662,7 @@ void PropagateDownloadFile::slotGetFinished()
 
         // Don't keep the temporary file if it is empty or we
         // used a bad range header or the file's not on the server anymore.
-        if (_tmpFile.size() == 0 || badRangeHeader || fileNotFound) {
+        if (_tmpFile.exists() && (_tmpFile.size() == 0 || badRangeHeader || fileNotFound)) {
             _tmpFile.close();
             FileSystem::remove(_tmpFile.fileName());
             propagator()->_journal->setDownloadInfo(_item->_file, SyncJournalDb::DownloadInfo());
@@ -968,6 +999,10 @@ void PropagateDownloadFile::downloadFinished()
         done(SyncFileItem::SoftError, error);
         return;
     }
+
+    // Make the file a hydrated placeholder if possible
+    propagator()->syncOptions()._vfs->convertToPlaceholder(fn, _item);
+
     FileSystem::setFileHidden(fn, false);
 
     // Maybe we downloaded a newer version of the file than we thought we would...
@@ -979,15 +1014,20 @@ void PropagateDownloadFile::downloadFinished()
     if (_conflictRecord.isValid())
         propagator()->_journal->setConflictRecord(_conflictRecord);
 
-    // If we downloaded something that used to be a virtual file,
-    // wipe the virtual file and its db entry now that we're done.
     if (_item->_type == ItemTypeVirtualFileDownload) {
-        auto virtualFile = propagator()->addVirtualFileSuffix(_item->_file);
-        auto fn = propagator()->getFilePath(virtualFile);
-        qCDebug(lcPropagateDownload) << "Download of previous virtual file finished" << fn;
-        QFile::remove(fn);
-        propagator()->_journal->deleteFileRecord(virtualFile);
+        // A downloaded virtual file becomes normal
         _item->_type = ItemTypeFile;
+
+        // If the virtual file used to have a different name and db
+        // entry, wipe both now.
+        auto vfs = propagator()->syncOptions()._vfs;
+        if (vfs && vfs->mode() == Vfs::WithSuffix) {
+            QString virtualFile = _item->_file + vfs->fileSuffix();
+            auto fn = propagator()->getFilePath(virtualFile);
+            qCDebug(lcPropagateDownload) << "Download of previous virtual file finished" << fn;
+            QFile::remove(fn);
+            propagator()->_journal->deleteFileRecord(virtualFile);
+        }
     }
 
     updateMetadata(isConflict);
